Remove dependency on meta crate from vex support
Also removes unneeded code and code cleans-up extraneous comments.
