+++ cranelift/assembler-x64/src/vex.rs
@@ -1,12 +1,38 @@
 //! Encoding logic for VEX instructions.
-//use super::evex::{Register, RegisterOrAmode};
-
 use super::rex;
 use super::XmmMem;
 use crate::api::{AsReg, CodeSink, KnownOffsetTable, Registers};
 use crate::mem::emit_modrm_sib_disp;
 use crate::Amode;
-use cranelift_assembler_x64_meta::dsl::{OpcodeMap, Vex};
+
+/// Allows using the same opcode byte in different "opcode maps" to allow for more instruction
+/// encodings. See appendix A in the Intel Software Developer's Manual, volume 2A, for more details.
+#[derive(PartialEq)]
+pub enum OpcodeMap {
+    None,
+    _0F,
+    _0F38,
+    _0F3A,
+}
+
+impl OpcodeMap {
+    /// Normally the opcode map is specified as bytes in the instruction, but some x64 encoding
+    /// formats pack this information as bits in a prefix (e.g. VEX / EVEX).
+    pub fn bits(&self) -> u8 {
+        match self {
+            OpcodeMap::None => 0b00,
+            OpcodeMap::_0F => 0b01,
+            OpcodeMap::_0F38 => 0b10,
+            OpcodeMap::_0F3A => 0b11,
+        }
+    }
+}
+
+impl Default for OpcodeMap {
+    fn default() -> Self {
+        Self::None
+    }
+}
 
 /// We may need to include one or more legacy prefix bytes before the REX prefix.  This enum
 /// covers only the small set of possibilities that we actually need.
@@ -79,110 +105,6 @@ impl<R: Registers> Default for VexInstruction<R> {
 }
 
 impl<R: Registers> VexInstruction<R> {
-    /// Construct a default VEX instruction.
-    pub fn new(vex: Vex) -> Self {
-        let mut vex_instruction = Self::default();
-        vex_instruction.opcode = vex.opcodes.primary;
-        vex_instruction
-    }
-
-    /// Set the length of the instruction.
-    #[inline(always)]
-    pub fn length(mut self, length: VexVectorLength) -> Self {
-        self.length = length;
-        self
-    }
-
-    /// Set the legacy prefix byte of the instruction: None | 66 | F2 | F3. VEX instructions
-    /// pack these into the prefix, not as separate bytes.
-    #[inline(always)]
-    pub fn prefix(mut self, prefix: LegacyPrefix) -> Self {
-        debug_assert!(
-            prefix == LegacyPrefix::None
-                || prefix == LegacyPrefix::_66
-                || prefix == LegacyPrefix::_F2
-                || prefix == LegacyPrefix::_F3
-        );
-
-        self.prefix = prefix;
-        self
-    }
-
-    /// Set the opcode map byte of the instruction: None | 0F | 0F38 | 0F3A. VEX instructions pack
-    /// these into the prefix, not as separate bytes.
-    #[inline(always)]
-    pub fn map(mut self, map: OpcodeMap) -> Self {
-        self.map = map;
-        self
-    }
-
-    /// Set the W bit, denoted by `.W1` or `.W0` in the instruction string.
-    /// Typically used to indicate an instruction using 64 bits of an operand (e.g.
-    /// 64 bit lanes). EVEX packs this bit in the EVEX prefix; previous encodings used the REX
-    /// prefix.
-    #[inline(always)]
-    pub fn w(mut self, w: bool) -> Self {
-        self.w = w;
-        self
-    }
-
-    /// Set the instruction opcode byte.
-    #[inline(always)]
-    pub fn opcode(mut self, opcode: u8) -> Self {
-        self.opcode = opcode;
-        self
-    }
-
-    /// Set the register to use for the `reg` bits; many instructions use this as the write operand.
-    #[inline(always)]
-    pub fn reg(mut self, reg: impl AsReg) -> Self {
-        self.reg = reg.enc();
-        self
-    }
-
-    /// Some instructions use the ModRM.reg field as an opcode extension. This is usually denoted by
-    /// a `/n` field in the manual.
-    #[inline(always)]
-    pub fn opcode_ext(mut self, n: u8) -> Self {
-        self.reg = n;
-        self
-    }
-
-    /// Set the register to use for the `rm` bits; many instructions use this
-    /// as the "read from register/memory" operand. Setting this affects both
-    /// the ModRM byte (`rm` section) and the VEX prefix (the extension bits
-    /// for register encodings > 8).
-    #[inline(always)]
-    pub fn rm(mut self, reg: XmmMem<R::ReadXmm, R::ReadGpr>) -> Self {
-        self.rm = reg.into();
-        self
-    }
-
-    /// Set the `vvvv` register; some instructions allow using this as a second, non-destructive
-    /// source register in 3-operand instructions (e.g. 2 read, 1 write).
-    #[inline(always)]
-    pub fn vvvv(mut self, reg: impl Into<u8>) -> Self {
-        self.vvvv = Some(reg.into());
-        self
-    }
-
-    /// Set the imm byte when used for a register. The reg bits are stored in `imm8[7:4]` with
-    /// the lower bits unused. Overrides a previously set [Self::imm] field.
-    #[inline(always)]
-    pub fn imm_reg(mut self, reg: impl AsReg) -> Self {
-        let reg: u8 = reg.enc();
-        self.imm = Some((reg & 0xf) << 4);
-        self
-    }
-
-    /// Set the imm byte.
-    /// Overrides a previously set [Self::imm_reg] field.
-    #[inline(always)]
-    pub fn imm(mut self, imm: u8) -> Self {
-        self.imm = Some(imm);
-        self
-    }
-
     /// The R bit in encoded format (inverted).
     #[inline(always)]
     fn r_bit(&self) -> u8 {
@@ -294,8 +216,6 @@ impl<R: Registers> VexInstruction<R> {
         match &self.rm {
             // Not all instructions use Reg as a reg, some use it as an extension
             // of the opcode.
-
-            //RegisterOrAmode::Register(reg) => {
             XmmMem::Xmm(xmm) => {
                 let rm: u8 = (*xmm).enc();
                 sink.put1(rex::encode_modrm(3, self.reg & 7, rm & 7));
