+++ cranelift/codegen/src/isa/x64/inst/external.rs
@@ -206,12 +206,10 @@ impl<'a, T: OperandVisitor> asm::RegisterVisitor<CraneliftRegisters> for Regallo
 
     fn write_xmm(&mut self, _reg: &mut Xmm) {
         todo!()
-        //self.collector.reg_reuse_def(reg, 0);
     }
 
     fn fixed_write_xmm(&mut self, _reg: &mut Xmm, _enc: u8) {
         todo!()
-        //self.collector.reg_reuse_def(reg, 0);
     }
 }
